LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/interaction/focus/tab-table-caption_caption-side=bottom-expected.txt

First tabbable span
Tabbable in cell
Tabbable in caption
Last tabbable span

FAIL Tab on <span tabindex="0">First tabbable span</span> should move focus to <span tabindex="0">Tabbable in cell<span></span></span> assert_equals: expected Element node <span tabindex="0">Tabbable in cell<span></span></span> but got Element node <span tabindex="0">First tabbable span</span>
FAIL Shift+Tab on <span tabindex="0">Tabbable in caption</span> should move focus to <span tabindex="0">Tabbable in cell<span></span></span> assert_equals: expected Element node <span tabindex="0">Tabbable in cell<span></span></span> but got Element node <span tabindex="0">Tabbable in caption</span>
FAIL Tab on <span tabindex="0">Tabbable in caption</span> should move focus to <span tabindex="0">Last tabbable span</span> assert_equals: expected Element node <span tabindex="0">Last tabbable span</span> but got Element node <span tabindex="0">Tabbable in caption</span>
FAIL Shift+Tab on <span tabindex="0">Tabbable in cell<span></span></span> should move focus to <span tabindex="0">First tabbable span</span> assert_equals: expected Element node <span tabindex="0">First tabbable span</span> but got Element node <span tabindex="0">Tabbable in cell<span></span></span>
FAIL Tab on <span tabindex="0">Tabbable in cell<span></span></span> should move focus to <span tabindex="0">Tabbable in caption</span> assert_equals: expected Element node <span tabindex="0">Tabbable in caption</span> but got Element node <span tabindex="0">Tabbable in cell<span></span></span>
FAIL Shift+Tab on <span tabindex="0">Last tabbable span</span> should move focus to <span tabindex="0">Tabbable in caption</span> assert_equals: expected Element node <span tabindex="0">Tabbable in caption</span> but got Element node <span tabindex="0">Last tabbable span</span>
